working but not stable

+++ philo/utils/ft_get_time.c
@@ -6,13 +6,13 @@
 /*   By: yel-moun <yel-moun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/31 19:07:18 by yel-moun          #+#    #+#             */
-/*   Updated: 2024/06/01 17:29:56 by yel-moun         ###   ########.fr       */
+/*   Updated: 2024/06/03 19:57:25 by yel-moun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../philosophers.h"
 
-long	ft_get_time()
+size_t	ft_get_time()
 {
 	struct timeval tv;
